MEL-25_p1_marketing_intel/api-gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.7
redis==5.0.1
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0 